nix cleanup

+++ shell.nix
@@ -9,8 +9,12 @@ mkShell rec {
     gnused
   ];
   shellHook = ''
-    cp pyrightconfig.json{,.bk}
-    VENV_NAME="$(echo ${pyEnv} | sed -E 's/\/nix\/store\/(.*)-env/\1/')"
-    sed -E -i "s/(\"venv\": \")\w+-python3-[23].[0-9].[0-9]-env/\1$VENV_NAME/" pyrightconfig.json
+    export PYTHONPATH="$(git rev-parse --show-toplevel):$PYTHONPATH"
+    if [ -f pyrightconfig.json ]; then
+      cp pyrightconfig.json{,.bk}
+      VENV_NAME="$(echo ${pyEnv} | sed -E 's/\/nix\/store\/(.*)-env/\1/')"
+      echo "detected pyrightconfig.json, patching venv with $VENV_NAME"
+      sed -E -i "s/(\"venv\": \")\w+-python3-[23].[0-9].[0-9]-env/\1$VENV_NAME/" pyrightconfig.json
+    fi
   '';
 }
